Code optimization 1
